+++ lib/dawn/core.rb
@@ -112,7 +112,7 @@ module Dawn
         fn = p + conf_name if p.start_with?('/')
         # if outside $HOME the config file must be hidden
         fn = File.expand_path(p) + '/.'+conf_name if ! p.start_with?('/')
-        $logger.debug("found a config file: " + fn) if File.exist?(fn)
+        debug_me("found a config file: " + fn) if File.exist?(fn)
         return fn if File.exist?(fn)
       end
 
@@ -132,7 +132,7 @@ module Dawn
       File.open(conf_path, 'w') do |f|
         rv = f.write(YAML.dump(conf))
       end
-      $logger.debug(conf_path)
+      debug_me(conf_path)
 
       conf_path
     end
@@ -140,7 +140,7 @@ module Dawn
     def self.read_conf(file=nil)
       conf = {:verbose=>false, :output=>"tabular", :mvc=>"", :gemfile_scan=>false, :gemfile_name=>"", :filename=>nil, :debug=>false, :exit_on_warn => false, :enabled_checks=> Dawn::Kb::BasicCheck::ALLOWED_FAMILIES}
       begin
-        $logger.debug("returning a default config") if file.nil? or ! File.exist?(file)
+        debug_me("returning a default config") if file.nil? or ! File.exist?(file)
         return conf if file.nil?
         file = file.chop if (not file.nil? and file.end_with? '/')
         return conf if ! File.exist?(file)