+++ lib/dawn/engine.rb
@@ -6,7 +6,6 @@ require 'socket'
 
 module Dawn
   module Engine
-    include Dawn::Utils
 
     attr_reader :target
     attr_reader :name
@@ -272,6 +271,8 @@ module Dawn
     # otherwise
     def apply(name)
 
+      telemetry
+
       # FIXME.20140325
       # Now if no checks are loaded because knowledge base was not previously called, apply and apply_all proudly refuse to run.
       # Reason is simple, load_knowledge_base now needs enabled check array
@@ -294,14 +295,14 @@ module Dawn
     end
 
     def have_a_telemetry_id?
-      $logger.debug ($telemetry_id != ""  and ! $telemetry_id.nil?)
+      debug_me ($telemetry_id != ""  and ! $telemetry_id.nil?)
       return ($telemetry_id != ""  and ! $telemetry_id.nil?)
       
     end
 
     def get_a_telemetry_id
       return "" if ($telemetry_url == "" or $telemetry_url.nil?)
-      $logger.debug("T: " + $telemetry_url)
+      debug_me("T: " + $telemetry_url)
 
       url = URI.parse($telemetry_url+"/new")
       res = Net::HTTP.get_response(url)
@@ -314,24 +315,25 @@ module Dawn
       unless have_a_telemetry_id?
         $telemetry_id = get_a_telemetry_id
         $config[:telemetry][:id] = $telemetry_id
-        $logger.debug($config)
-        $logger.debug("saving config to " + $config_name)
+        debug_me($config)
+        debug_me("saving config to " + $config_name)
         File.open($config_name, 'w') { |f| f.write $config.to_yaml }
       end
 
-      $logger.debug("Telemetry ID is: " + $telemetry_id)
+      debug_me("Telemetry ID is: " + $telemetry_id)
       
       uri=URI.parse($telemetry_url+"/"+$telemetry_id)
       header = {'Content-Type': 'text/json'}
       tele = { "kb_version" => Dawn::KnowledgeBase::VERSION , 
-               "ip" => Socket.ip_address_list.detect{|intf| intf.ipv4_private?}.ip_address
+               "ip" => Socket.ip_address_list.detect{|intf| intf.ipv4_private?}.ip_address, 
+               "message"=> Dawn::KnowledgeBase
             }
       http = Net::HTTP.new(uri.host, uri.port)
       request = Net::HTTP::Post.new(uri.request_uri, header)
       request.body = tele.to_json
 
       response=http.request(request)
-      $logger.debug(response.inspect)
+      debug_me(response.inspect)
 
       return true