+++ Changelog.md
@@ -15,6 +15,8 @@ _latest update: mer 28 nov 2018, 11.03.53, CET_
 * Added a new debug\_verbosely API for engines and checks
 * Removed rake osvdb[name] and rake cve[name] tasks
 * Adding telemetry
+* Dawn::Utils include refactory. Now it's available application wide
+* debug information refactory. 
 
 ## Version 1.6.9 - codename: Tow Mater (2018-11-28)